requirements.txt
requests==2.23.0
python_dotenv==0.13.0